backend/requirements.txt
asgiref==3.9.1
cffi==2.0.0
cryptography==45.0.7
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
mysqlclient==2.2.7
pycparser==2.23
PyMySQL==1.1.2
sqlparse==0.5.3
tzdata==2025.2
